Add kbn/test-es-server package (#259064)
## Summary

Introduces `@kbn/test-es-server` as a focused package for
starting/managing Elasticsearch in tests (cluster helpers, ES test
config, createEsClientForTesting, cleanup, CI parallel prefix).
`@kbn/test` and `@kbn/scout` depend on it and keep re-exporting the same
public surface where needed so existing imports keep working.

FTR-specific wiring that needs the FTR Config type stays in `@kbn/test`
(e.g. `createEsClientForFtrConfig` in `ftr_es_client.ts`).

`@kbn/evals` now imports `createEsClientForTesting` from
`@kbn/test-es-server` instead of `@kbn/test`, aligning it with the
Scout-side ES client path and trimming the implied dependency on the
full test harness for that call.

### Why
- Clearer boundaries: ES test-server concerns live in one place;
@kbn/test stays the broader FTR/harness package.
- Easier evolution: Scout and other tooling can depend on ES test
infrastructure without pulling all of @kbn/test when they only need ES
helpers.

### Scope
- New `@kbn/test-es-server` package and repo wiring (paths, links,
moon/tsconfig as required).
- `@kbn/test` / `@kbn/scout`: delegate ES cluster/client/config/cleanup
(and related imports) to the new package; FTR helpers remain in
`@kbn/test.`
- `@kbn/evals`: switch `createEsClientForTesting` to
`@kbn/test-es-server` and update package references.

---------

(cherry picked from commit 592399bbf8fbb25af3b4e4aab580e1f29c546324)

# Conflicts:
#	x-pack/platform/packages/shared/kbn-evals/src/cli/commands/compare.ts
#	x-pack/platform/packages/shared/kbn-evals/tsconfig.json
