+++ packages/kbn-repo-source-classifier/src/repo_source_classifier.ts
@@ -162,7 +162,11 @@ export class RepoSourceClassifier {
 
     const { pkgId, rel } = pkgInfo;
 
-    if (pkgId === '@kbn/test' || pkgId === '@kbn/test-subj-selector') {
+    if (
+      pkgId === '@kbn/test' ||
+      pkgId === '@kbn/test-es-server' ||
+      pkgId === '@kbn/test-subj-selector'
+    ) {
       return 'common package';
     }
