+++ src/platform/packages/shared/kbn-scout/src/common/services/clients.ts
@@ -8,7 +8,7 @@
  */
 
 import { KbnClient } from '@kbn/kbn-client';
-import { createEsClientForTesting } from '@kbn/test';
+import { createEsClientForTesting } from '@kbn/test-es-server';
 import type { ScoutLogger } from './logger';
 import type { ScoutTestConfig, EsClient } from '../../types';
