+++ src/platform/packages/shared/kbn-test-es-server/index.ts
@@ -7,12 +7,16 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { createTestEsCluster } from './test_es_cluster';
-export type { CreateTestEsClusterOptions, EsTestCluster, ICluster } from './test_es_cluster';
-export { esTestConfig } from './es_test_config';
+export { CI_PARALLEL_PROCESS_PREFIX } from './src/ci_parallel_process_prefix';
+export { cleanupElasticsearch } from './src/cleanup_elasticsearch';
 export {
   createEsClientForTesting,
-  createEsClientForFtrConfig,
-  createRemoteEsClientForFtrConfig,
-} from './es_client_for_testing';
-export type { EsClientForTestingOptions } from './es_client_for_testing';
+  type EsClientForTestingOptions,
+} from './src/es_client_for_testing';
+export { esTestConfig } from './src/es_test_config';
+export {
+  createTestEsCluster,
+  type CreateTestEsClusterOptions,
+  type EsTestCluster,
+  type ICluster,
+} from './src/test_es_cluster';
