+++ src/platform/packages/shared/kbn-test-es-server/kibana.jsonc
@@ -1,0 +1,11 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/test-es-server",
+  "owner": [
+    "@elastic/kibana-operations",
+    "@elastic/appex-qa"
+  ],
+  "group": "platform",
+  "visibility": "shared",
+  "devOnly": true
+}
