+++ src/platform/packages/shared/kbn-test-es-server/moon.yml
@@ -1,47 +1,28 @@
 # This file is generated by the @kbn/moon package. Any manual edits will be erased!
 #  To extend this, write your extensions/overrides to 'moon.extend.yml'
-#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/scout'
+#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/test-es-server'
 
 $schema: https://moonrepo.dev/schemas/project.json
-id: '@kbn/scout'
+id: '@kbn/test-es-server'
 layer: unknown
 owners:
-  defaultOwner: '@elastic/appex-qa'
+  defaultOwner: '@elastic/kibana-operations'
 toolchains:
   default: node
 language: typescript
 project:
-  title: '@kbn/scout'
-  description: Moon project for @kbn/scout
+  title: '@kbn/test-es-server'
+  description: Moon project for @kbn/test-es-server
   channel: ''
-  owner: '@elastic/appex-qa'
-  sourceRoot: src/platform/packages/shared/kbn-scout
+  owner: '@elastic/kibana-operations'
+  sourceRoot: src/platform/packages/shared/kbn-test-es-server
 dependsOn:
-  - '@kbn/tooling-log'
-  - '@kbn/dev-cli-runner'
-  - '@kbn/dev-cli-errors'
-  - '@kbn/ci-stats-reporter'
-  - '@kbn/repo-info'
-  - '@kbn/repo-packages'
-  - '@kbn/es'
-  - '@kbn/dev-proc-runner'
-  - '@kbn/kbn-client'
-  - '@kbn/test-es-server'
-  - '@kbn/test'
-  - '@kbn/es-archiver'
   - '@kbn/dev-utils'
-  - '@kbn/mock-idp-utils'
-  - '@kbn/test-subj-selector'
-  - '@kbn/scout-info'
-  - '@kbn/scout-reporting'
-  - '@kbn/synthtrace'
-  - '@kbn/synthtrace-client'
-  - '@kbn/zod'
-  - '@kbn/core-http-common'
-  - '@kbn/axe-config'
-  - '@kbn/peggy'
+  - '@kbn/es'
+  - '@kbn/repo-info'
+  - '@kbn/tooling-log'
 tags:
-  - test-helper
+  - shared-common
   - package
   - dev
   - group-platform
@@ -50,6 +31,7 @@ tags:
 fileGroups:
   src:
     - '**/*.ts'
+    - '**/*.js'
     - '!target/**/*'
 tasks:
   jest:
