+++ src/platform/packages/shared/kbn-test-es-server/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@kbn/test-es-server",
+  "author": "Operations",
+  "version": "1.0.0",
+  "private": true,
+  "license": "Elastic License 2.0 OR AGPL-3.0-only OR SSPL-1.0"
+}
