+++ src/platform/packages/shared/kbn-test-es-server/src/test_es_cluster.ts
@@ -22,7 +22,7 @@ import { REPO_ROOT } from '@kbn/repo-info';
 import type { ArtifactLicense } from '@kbn/es';
 import type { ServerlessOptions } from '@kbn/es/src/utils';
 import { getFips } from 'crypto';
-import { CI_PARALLEL_PROCESS_PREFIX } from '../ci_parallel_process_prefix';
+import { CI_PARALLEL_PROCESS_PREFIX } from './ci_parallel_process_prefix';
 import { esTestConfig } from './es_test_config';
 
 interface TestEsClusterNodesOptions {
