+++ src/platform/packages/shared/kbn-test-es-server/tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "extends": "@kbn/tsconfig-base/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "target/types",
+    "stripInternal": true,
+    "types": ["jest", "node"]
+  },
+  "include": ["**/*.ts", "**/*.js"],
+  "exclude": ["target/**/*"],
+  "kbn_references": [
+    "@kbn/dev-utils",
+    "@kbn/es",
+    "@kbn/repo-info",
+    "@kbn/tooling-log"
+  ]
+}
