+++ src/platform/packages/shared/kbn-test/index.ts
@@ -37,14 +37,9 @@ export type {
   EsTestCluster,
   ICluster,
   EsClientForTestingOptions,
-} from './src/es';
-export {
-  esTestConfig,
-  createTestEsCluster,
-  createEsClientForTesting,
-  createEsClientForFtrConfig,
-  createRemoteEsClientForFtrConfig,
-} from './src/es';
+} from '@kbn/test-es-server';
+export { esTestConfig, createTestEsCluster, createEsClientForTesting } from '@kbn/test-es-server';
+export { createEsClientForFtrConfig, createRemoteEsClientForFtrConfig } from './src/ftr_es_client';
 
 export { kbnTestConfig } from './kbn_test_config';
 export type { UrlParts } from './kbn_test_config';
@@ -60,7 +55,7 @@ export {
 // @internal
 export { setupJUnitReportGeneration, escapeCdata } from './src/mocha';
 
-export { CI_PARALLEL_PROCESS_PREFIX } from './src/ci_parallel_process_prefix';
+export { CI_PARALLEL_PROCESS_PREFIX } from '@kbn/test-es-server';
 
 export * from './src/functional_test_runner';
