+++ src/platform/packages/shared/kbn-test/moon.yml
@@ -18,8 +18,8 @@ project:
   sourceRoot: src/platform/packages/shared/kbn-test
 dependsOn:
   - '@kbn/kbn-client'
-  - '@kbn/dev-utils'
   - '@kbn/std'
+  - '@kbn/test-es-server'
   - '@kbn/tooling-log'
   - '@kbn/get-repo-files'
   - '@kbn/peggy'
