+++ src/platform/packages/shared/kbn-test/tsconfig.json
@@ -3,25 +3,14 @@
   "compilerOptions": {
     "outDir": "target/types",
     "stripInternal": true,
-    "types": [
-      "jest",
-      "node"
-    ]
+    "types": ["jest", "node"]
   },
-  "include": [
-    "**/*.ts",
-    "**/*.js",
-    "src/jest/jest_flags.json",
-  ],
-  "exclude": [
-    "types/**/*",
-    "**/__fixtures__/**/*",
-    "target/**/*",
-  ],
+  "include": ["**/*.ts", "**/*.js", "src/jest/jest_flags.json"],
+  "exclude": ["types/**/*", "**/__fixtures__/**/*", "target/**/*"],
   "kbn_references": [
     "@kbn/kbn-client",
-    "@kbn/dev-utils",
     "@kbn/std",
+    "@kbn/test-es-server",
     "@kbn/tooling-log",
     "@kbn/get-repo-files",
     "@kbn/peggy",
@@ -42,6 +31,6 @@
     "@kbn/react-mute-legacy-root-warning",
     "@kbn/babel-preset",
     "@kbn/dot-text",
-    "@kbn/apm-utils",
+    "@kbn/apm-utils"
   ]
 }
