+++ x-pack/platform/packages/shared/kbn-evals/moon.yml
@@ -31,7 +31,7 @@ dependsOn:
   - '@kbn/repo-info'
   - '@kbn/std'
   - '@kbn/kbn-client'
-  - '@kbn/test'
+  - '@kbn/test-es-server'
   - '@kbn/inference-prompt-utils'
   - '@kbn/tracing-utils'
 tags:
